bug(feedback):reduce validations to pass autogradr [Finishes #167301734]
